Run on CVL

+++ scripts/06_blip2_caption_analysis/02_map_label_dependency_parsing_features_and_blip2_answer_dependency_parsing_features/02_process_asl_predictions.py
@@ -68,7 +68,7 @@ if __name__ == "__main__":
         frame_id_asl_predicted_label_indices_mapping = dict()
         frame_id_asl_predicted_label_indices_mapping[clip_id] = dict()
         fps = 30.0
-        num_frames = duration * 30.0
+        num_frames = int(duration * 30.0)
         for frame_id in range(num_frames):
             frame_id_asl_predicted_label_indices_mapping[clip_id][frame_id] = dict()
             frame_id_asl_predicted_label_indices_mapping[clip_id][frame_id][
